uint64 -> int64 conversion

+++ src/types.jl
@@ -434,7 +434,7 @@ function Base.unsafe_wrap(v::UA_Variant)
     values = GC.@preserve data unsafe_wrap(Array, data, unsafe_size(v))
     values_row_major = reshape(values, unsafe_size(v))
     @show v, v.arrayDimensionsSize
-    return permutedims(values_row_major, reverse(1:(v.arrayDimensionsSize))) # To column major format; TODO: Which permutation is right? TODO: can make allocation free using PermutedDimsArray?
+    return permutedims(values_row_major, reverse(1:(Int64(v.arrayDimensionsSize)))) # To column major format; TODO: Which permutation is right? TODO: can make allocation free using PermutedDimsArray?
 end
 
 Base.unsafe_wrap(p::Ref{UA_Variant}) = unsafe_wrap(unsafe_load(p))
